fix some naming issues.
