+++ src/components/RentSearch.js
@@ -1,0 +1,5 @@
+import React from 'react'
+
+export default function RentSearch() {
+    return (<div></div>)
+}
